server/api/requirements.txt
Django >=3, < 4

# System
Pillow 

# Database
psycopg2

# Security
argon2-cffi
django-filter
django-axes
django-guardian
django-oauth-toolkit
django-cors-middleware

# GraphQL
django-graphql-jwt
graphene-django
